refactor : step3 로또 객체 domain package로 이동
