+++ src/test/java/step3/LottoTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import step3.domain.Lotto;
 
 @DisplayName("Lotto 클래스")
 public class LottoTest {
